Create a List of Opportunity Accounts
By joining TDlinx Master & VIPSAM2 data sources we will be able to provide a list of "opportunity accounts". 
Opportunity accounts are stores that we have not sold a product in within the previous 12 months.

For the Proof-of-Concept  this will be built out at the [tdlinx_number and brand] level and once all parties agree on how we will proceed with the Opportunity Accounts, then we will build out the data at the [tdlinx, brand, sub-brand, bottle size, and SKU] level.