After geocode_tdlinx_master.Rmd runs, fill in missing coords of stores
geocode_tdlinx_master.Rmd will attempt to create lat/long coordinates for all stores (tdlinx numbers). 

The purpose of this script is to populate any missing coordinates with that of the store's zip code or city (if zip code does not work). 

This script will need to be updated with correct file names (to read in) once the geocode_tdlinx_master.Rmd outputs the geocoded_address.csv file.