Moving geocoding files into their own folder
Don't want to confuse anyone with the extra files, so moving them to a sub-folder. Files within the main folder remain there because they are the only files need to achieve the primary goal of the project. 

The geocoding part of the project is secondary, and falls into the "would be nice to have" category.